# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API para gestionar tareas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/fees": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Lista todas las cuotas",
                "operationId": "6968c794805ccbda232e2c34d376eeed",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar cuotas por nombre de cliente.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de cuotas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Fee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Fees"
                ],
                "summary": "Almacena una nueva cuota extraordinaria",
                "operationId": "567ac89b621efb68d966cb786c2ca0b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreNewExtraordinaryFeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de cuotas con mensaje de éxito."
                    }
                }
            }
        },
...
        "/fees/{id}/send-invoice": {
            "post": {
                "tags": [
                    "Fees"
                ],
                "summary": "Envía la factura de una cuota por correo electrónico",
                "operationId": "5dec78fc08233324ad373a74e5e5a782",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuota.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Factura enviada correctamente."
                    }
                }
            }
        },
        "/fees/{id}": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Muestra los detalles de una cuota",
                "operationId": "397b6250baa868a6292741e9180da60f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuota.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la cuota.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fee"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Fees"
                ],
                "summary": "Elimina una cuota",
                "operationId": "f0855d57e721fa7d5bf793c774314d18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuota a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de cuotas con mensaje de éxito."
                    }
                }
            }
        },
        "/fees/{id}/edit": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Muestra el formulario para editar una cuota",
                "operationId": "4737667ffd39c788fa72235ea6d8146b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuota a editar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario de edición de cuota."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreNewExtraordinaryFeeRequest": {
                "title": "Store New Extraordinary Fee Request",
                "description": "Request para almacenar una nueva cuota extraordinaria.",
                "required": [
                    "cif",
                    "concept",
                    "amount",
                    "notes"
                ],
                "properties": {
                    "cif": {
                        "description": "CIF del cliente.",
                        "type": "string"
                    },
                    "concept": {
                        "description": "Concepto de la cuota extraordinaria.",
                        "type": "string"
                    },
                    "amount": {
                        "description": "Monto de la cuota extraordinaria.",
                        "type": "number",
                        "format": "float"
                    },
                    "notes": {
                        "description": "Notas adicionales.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Fee": {
                "title": "Fee",
                "description": "Modelo de cuota.",
                "properties": {
                    "id": {
                        "description": "ID de la cuota.",
                        "type": "integer"
                    },
                    "cif": {
                        "description": "CIF del cliente.",
                        "type": "string"
                    },
                    "concept": {
                        "description": "Concepto de la cuota.",
                        "type": "string"
                    },
                    "issue_date": {
                        "description": "Fecha de emisión de la cuota.",
                        "type": "string",
                        "format": "date"
                    },
                    "amount": {
                        "description": "Importe de la cuota.",
                        "type": "number",
                        "format": "decimal"
                    },
                    "passed": {
                        "description": "Indica si la cuota ha sido pasada.",
                        "type": "boolean"
                    },
                    "payment_date": {
                        "description": "Fecha de pago de la cuota.",
                        "type": "string",
                        "format": "date"
                    },
                    "notes": {
                        "description": "Notas adicionales sobre la cuota.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}